% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-wrapper.R
\name{datagoodr}
\alias{datagoodr}
\title{Generate DGF and RG}
\usage{
datagoodr(
  wd = getwd(),
  folder.name = NULL,
  path.raw.data,
  create.dgf.params = NULL,
  rg.name = "research-guide"
)
}
\arguments{
\item{wd}{charterer string of project output location. Default is the current working directory.}

\item{folder.name}{character string of folder project output will be put into. Defaut is "datagoodr-TODAYSDATE".}

\item{path.raw.data}{character string of absolute path to raw dataset that is to be documented. File extension can be .csv or .xlsx. See data-dev/demo-data-small.csv for an example.}

\item{create.dgf.params}{list of function arguments for creating the dgf. See \link{create_dgf} for available function arguments (df argument is not needed). Defaul is NULL which uses the default arguments for \link{create_dgf}.}

\item{rg.name}{character string to name the final research guide output. Default is "research-guide".}
}
\value{
This function outputs entire datagoodr workflow into the `wd` directory.
}
\description{
Run the entire datagoodr workflow
}
\details{
Run the entire datagoodr workflow.....
}
\examples{
\dontrun{
## Testing datagoor funciton
library(datagoodr )
datagoodr(
  wd = "~/Desktop",
  folder.name = NULL,
  location.data.raw <- paste0(getwd(), "/data-dev/DEMO-DATA-SMALL.csv"),
  create.dgf.params = list(use.df.types = FALSE,
                           guess.factors = TRUE,
                           guess.dates = FALSE),
  rg.name = "test-research-guide"
)
}

}
