% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-data.R
\name{dgf}
\alias{dgf}
\title{DGF Default}
\format{
Result of \link{create_dgf} with \link{example.df} data.
}
\description{
An example of a generated DGF with all default options.
}
\details{
This data set was generated using the code in the example section of this help file.
}
\examples{
\dontrun{
data(df)
create_dgf(df, file = "dgf")
}
}
