% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{dollarize}
\alias{dollarize}
\title{Format Numeric Values as Dollar Amounts}
\usage{
dollarize(x)
}
\arguments{
\item{x}{A numeric vector to be formatted.}
}
\value{
A character vector with the formatted dollar amounts.
}
\description{
This function formats numeric values as dollar amounts by rounding them to the nearest whole number,
adding commas as thousand separators, and prefixing them with a dollar sign.
}
\examples{
dollarize(1234567.89) # Returns "$1,234,568"
dollarize(c(1000, 2500.75, 999999.99)) # Returns c("$1,000", "$2,501", "$1,000,000")

}
