% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-data.R
\name{example.df}
\alias{example.df}
\title{Example IRS990 Data Set}
\format{
A data frame with 25,000 rows and 27 variables:
\describe{
  \item{F9_00_ORG_NAME_L1}{Organization name line 1}
  \item{F0_00_ORG_CONTACT}{Contact person (from IRS files)}
  \item{F9_00_ORG_ADDR_L1}{Organization street address line 1}
  \item{F9_00_ORG_ADDR_CITY}{Organization city}
  \item{EIN}{EIN}
  \item{SUBSECCD}{IRS subsection code}
  \item{BMF_ACTIV1}{IRS Activity Code 1}
  \item{NTMAJ12}{NTEE major group (12)}
  \item{NTEE1}{NTEE major group}
  \item{NTEEFINAL}{NA}
  \item{NTEESRC}{NA}
  \item{DEDUCTCD}{IRS Deductibility code}
  \item{OUTREAS}{Reason why out of scope}
  \item{F9_05_UBIZ_IMCOME_OVER_LIMIT_X}{Had unrelated business gross income of \$1,000 or more}
  \item{OUTNCCS}{Out of Scope flag}
  \item{COUNTY_FIPS}{State + County FIPS code}
  \item{CEO_CENSUSTRACT}{Census tract}
  \item{F9_00_TAX_PERIOD_END_DATE}{Tax period end date}
  \item{F9_00_TAX_PERIOD_END_DATE_PY}{Tax period end date - prior year}
  \item{F9_00_TAX_PERIOD_BEGIN_DATE}{Tax period begin date}
  \item{F9_00_TAX_ACCPER}{Tax period end date}
  \item{F9_08_REV_TOT_TOT}{Total revenue - total}
  \item{F9_10_ASSET_TOT_BOY}{Total assets - beginning of year}
  \item{F9_10_ASSET_TOT_EOY}{Total assets - end of year}
  \item{F9_10_NAFB_TOT_BOY}{Net assets or fund balances - beginning of year}
  \item{F9_09_EXP_TOT_TOT}{Total functional expenses - total expenses}
  \item{F9_01_EXP_TOT_PY}{Total expenses - prior year}
}
}
\description{
A toy example of IRS990 data to be documented with datagoodr
}
\details{
See \url{https://nonprofit-open-data-collective.github.io/irs990efile/data-dictionary/data-dictionary.html} for full data dictionary.
}
