% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{get_dupes}
\alias{get_dupes}
\title{Identify and Print Duplicate Variables in a Data Frame}
\usage{
get_dupes(df, vhash)
}
\arguments{
\item{df}{A data frame containing the variables to be checked for raw_duplicated.}

\item{vhash}{A vector of hash values corresponding to each variable in the data frame.
This is typically computed using a hashing function like `rlang::hash()`.}
}
\value{
A logical vector indicating which variables are raw_duplicated.
}
\description{
This function checks for duplicate variables in a data frame based on a provided hash vector.
It prints out the names of duplicate variables and returns a logical vector indicating which
rows are raw_duplicated.
}
\examples{
\dontrun{
vhash <- sapply(df, rlang::hash)  # Generate hash values for variables
get_dupes(df, vhash)  # Identify and print duplicate variables
}

}
