% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-03-get-stats.R
\name{get_properties}
\alias{get_properties}
\title{Generate Table of Properties}
\usage{
get_properties(VNAME, df)
}
\arguments{
\item{VNAME}{A character string specifying the name of the variable in the data frame. (From var_name column of the DGF)}

\item{df}{A data frame containing the variable to be visualized.}
}
\value{
Data table as JSON text string needed to store in the DGF to later generate the properties table when rendering the RG.
}
\description{
Generates needed properties for rg_properties column of the DGF. Used inside R/01-00-CREATE-DGF.R
}
\details{
properties include number of rows and distinct values, most common value,
number of zero, empty cells, missing/NA cells, and cell with infinite values
}
