% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-codes.R
\name{is_ca_postal_code}
\alias{is_ca_postal_code}
\title{Is it a Canadian postal code (or FSA)?}
\usage{
is_ca_postal_code(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a 3-char FSA (`V5K`) or full `A1A 1A1`.
}
\description{
Is it a Canadian postal code (or FSA)?
}
\examples{
is_ca_postal_code(c("V5K", "K1A 0B1", "12345", NA))
}
\seealso{
Other code detectors: 
\code{\link{is_country_code}()},
\code{\link{is_http_status}()},
\code{\link{is_roman}()},
\code{\link{is_uk_postcode}()},
\code{\link{is_unix_time}()},
\code{\link{is_us_state}()},
\code{\link{is_zip_code}()}
}
\concept{code detectors}
