% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-geography.R
\name{is_cbsa}
\alias{is_cbsa}
\title{Is it a CBSA code?}
\usage{
is_cbsa(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a 5-digit Core Based Statistical Area
  code in the Census delineation (`.cbsa` lookup).
}
\description{
Is it a CBSA code?
}
\examples{
is_cbsa(c("10180", "35620", "99999", NA))
}
\seealso{
Other geography detectors: 
\code{\link{is_block_fips}()},
\code{\link{is_block_group_fips}()},
\code{\link{is_census_region}()},
\code{\link{is_county_fips}()},
\code{\link{is_csa}()},
\code{\link{is_metro_area_name}()},
\code{\link{is_nces_locale}()},
\code{\link{is_state_fips}()},
\code{\link{is_tract_fips}()},
\code{\link{is_urban_rural}()}
}
\concept{geography detectors}
