% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-geography.R
\name{is_census_region}
\alias{is_census_region}
\title{Is it a US census region or division?}
\usage{
is_census_region(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a census region (Northeast/Midwest/South/
  West) or one of the nine census divisions.
}
\description{
Is it a US census region or division?
}
\examples{
is_census_region(c("Northeast", "Pacific", "banana", NA))
}
\seealso{
Other geography detectors: 
\code{\link{is_block_fips}()},
\code{\link{is_block_group_fips}()},
\code{\link{is_cbsa}()},
\code{\link{is_county_fips}()},
\code{\link{is_csa}()},
\code{\link{is_metro_area_name}()},
\code{\link{is_nces_locale}()},
\code{\link{is_state_fips}()},
\code{\link{is_tract_fips}()},
\code{\link{is_urban_rural}()}
}
\concept{geography detectors}
