% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-gov.R
\name{is_chinese_resident_id}
\alias{is_chinese_resident_id}
\title{Is it a Chinese resident identity card number?}
\usage{
is_chinese_resident_id(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for an 18-character ID whose ISO 7064
  mod-11-2 check character is valid.
}
\description{
Is it a Chinese resident identity card number?
}
\examples{
is_chinese_resident_id(c("141002197808194474", NA))
}
\seealso{
Other gov detectors: 
\code{\link{is_dea}()},
\code{\link{is_ein}()},
\code{\link{is_naics}()},
\code{\link{is_ssn}()},
\code{\link{is_uei}()}
}
\concept{gov detectors}
