% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-address.R
\name{is_city_name}
\alias{is_city_name}
\title{Is it a US city / place name?}
\usage{
is_city_name(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a name in the normalized `.us_cities`
  lookup (case/punctuation-insensitive). Coverage is the ~1,000 larger US
  cities (maps::us.cities); smaller towns will miss.
}
\description{
Is it a US city / place name?
}
\examples{
is_city_name(c("Chicago", "New Orleans", "Nowheresville", NA))
}
\seealso{
Other address detectors: 
\code{\link{is_address}()},
\code{\link{is_city_state}()},
\code{\link{is_po_box}()},
\code{\link{is_state_zip}()},
\code{\link{is_street_address}()}
}
\concept{address detectors}
