% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-address.R
\name{is_city_state}
\alias{is_city_state}
\title{Is it a "City, State" value?}
\usage{
is_city_state(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for `<anything>, <state abbrev or name>`.
}
\description{
Is it a "City, State" value?
}
\examples{
is_city_state(c("Bloomington, IN", "Bloomington, Indiana", "just text", NA))
}
\seealso{
Other address detectors: 
\code{\link{is_address}()},
\code{\link{is_city_name}()},
\code{\link{is_po_box}()},
\code{\link{is_state_zip}()},
\code{\link{is_street_address}()}
}
\concept{address detectors}
