% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-geography.R
\name{is_county_fips}
\alias{is_county_fips}
\title{Is it a county FIPS code?}
\usage{
is_county_fips(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a 5-digit code with a valid state prefix.
}
\description{
Is it a county FIPS code?
}
\examples{
is_county_fips(c("06037", "36061", "99001", NA))
}
\seealso{
Other geography detectors: 
\code{\link{is_block_fips}()},
\code{\link{is_block_group_fips}()},
\code{\link{is_cbsa}()},
\code{\link{is_census_region}()},
\code{\link{is_csa}()},
\code{\link{is_metro_area_name}()},
\code{\link{is_nces_locale}()},
\code{\link{is_state_fips}()},
\code{\link{is_tract_fips}()},
\code{\link{is_urban_rural}()}
}
\concept{geography detectors}
