% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-gov.R
\name{is_dea}
\alias{is_dea}
\title{Is it a DEA registration number?}
\usage{
is_dea(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for two letters + seven digits whose DEA
  check digit is valid.
}
\description{
Is it a DEA registration number?
}
\examples{
is_dea(c("AP5836727", "AP5836728", NA))
}
\seealso{
Other gov detectors: 
\code{\link{is_chinese_resident_id}()},
\code{\link{is_ein}()},
\code{\link{is_naics}()},
\code{\link{is_ssn}()},
\code{\link{is_uei}()}
}
\concept{gov detectors}
