% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-research-id.R
\name{is_doi}
\alias{is_doi}
\title{Is it a DOI?}
\usage{
is_doi(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a `10.<registrant>/<suffix>` DOI.
}
\description{
Is it a DOI?
}
\examples{
is_doi(c("10.1109/ICDCSW.2017.43", "not-a-doi", NA))
}
\seealso{
Other research-id detectors: 
\code{\link{is_isbn}()},
\code{\link{is_issn}()},
\code{\link{is_npi}()},
\code{\link{is_orcid}()}
}
\concept{research-id detectors}
