% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-hash.R
\name{is_guid}
\alias{is_guid}
\title{Is it a GUID / UUID?}
\usage{
is_guid(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for the `8-4-4-4-12` hex UUID form.
}
\description{
Is it a GUID / UUID?
}
\examples{
is_guid(c("f2a59360-1d1a-49e8-9a2f-f9cdf0d5a811", NA))
}
\seealso{
Other hash detectors: 
\code{\link{is_md5}()},
\code{\link{is_sha1}()},
\code{\link{is_sha256}()}
}
\concept{hash detectors}
