% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-color.R
\name{is_hex_color}
\alias{is_hex_color}
\title{Is it a hex color?}
\usage{
is_hex_color(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for `#RGB` or `#RRGGBB`.
}
\description{
Is it a hex color?
}
\examples{
is_hex_color(c("#4d9cf8", "#fff", "4d9cf8", NA))
}
\seealso{
Other color detectors: 
\code{\link{is_cmyk_color}()},
\code{\link{is_hsl_color}()},
\code{\link{is_rgb_color}()}
}
\concept{color detectors}
