% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-codes.R
\name{is_http_status}
\alias{is_http_status}
\title{Is it an HTTP status code?}
\usage{
is_http_status(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a three-digit code in 100-599.
}
\description{
Is it an HTTP status code?
}
\examples{
is_http_status(c("200", "404", "099", "600", NA))
}
\seealso{
Other code detectors: 
\code{\link{is_ca_postal_code}()},
\code{\link{is_country_code}()},
\code{\link{is_roman}()},
\code{\link{is_uk_postcode}()},
\code{\link{is_unix_time}()},
\code{\link{is_us_state}()},
\code{\link{is_zip_code}()}
}
\concept{code detectors}
