% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-scibio.R
\name{is_icd9_code}
\alias{is_icd9_code}
\title{Is it an ICD-9 code?}
\usage{
is_icd9_code(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a numeric or E/V ICD-9-CM code.
}
\description{
Is it an ICD-9 code?
}
\examples{
is_icd9_code(c("078.11", "V70.0", "E880.9", NA))
}
\seealso{
Other scibio detectors: 
\code{\link{is_asin}()},
\code{\link{is_bibcode}()},
\code{\link{is_ensembl_id}()},
\code{\link{is_icd10_code}()},
\code{\link{is_snp_id}()},
\code{\link{is_uniprot_id}()}
}
\concept{scibio detectors}
