% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-phone.R
\name{is_intl_phone}
\alias{is_intl_phone}
\title{Is it an international (+ prefixed) phone number?}
\usage{
is_intl_phone(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a `+`-prefixed number with 8-15 digits
  and common separators/parentheses.
}
\description{
Is it an international (+ prefixed) phone number?
}
\examples{
is_intl_phone(c("+1-4250013981", "+61 3 6216 0864", "4250013981", NA))
}
\seealso{
Other phone detectors: 
\code{\link{is_msisdn}()},
\code{\link{is_phone}()}
}
\concept{phone detectors}
