% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-web.R
\name{is_ipv6}
\alias{is_ipv6}
\title{Is it an IPv6 address?}
\usage{
is_ipv6(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a full or `::`-compressed IPv6 address.
}
\description{
Is it an IPv6 address?
}
\examples{
is_ipv6(c("2001:0db8:85a3:0000:0000:8a2e:0370:7334", "1:2:3", NA))
}
\seealso{
Other web detectors: 
\code{\link{is_email}()},
\code{\link{is_ipv4}()},
\code{\link{is_mac}()},
\code{\link{is_url}()}
}
\concept{web detectors}
