% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-web.R
\name{is_mac}
\alias{is_mac}
\title{Is it a MAC address?}
\usage{
is_mac(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for six `:`- or `-`-separated hex octets.
}
\description{
Is it a MAC address?
}
\examples{
is_mac(c("00:1B:44:11:3A:B7", "40-1F-3C-B6-3C-BC", NA))
}
\seealso{
Other web detectors: 
\code{\link{is_email}()},
\code{\link{is_ipv4}()},
\code{\link{is_ipv6}()},
\code{\link{is_url}()}
}
\concept{web detectors}
