% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-hash.R
\name{is_md5}
\alias{is_md5}
\title{Is it an MD5 digest?}
\usage{
is_md5(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a 32-character hex string.
}
\description{
Is it an MD5 digest?
}
\examples{
is_md5(c("d41d8cd98f00b204e9800998ecf8427e", "abc", NA))
}
\seealso{
Other hash detectors: 
\code{\link{is_guid}()},
\code{\link{is_sha1}()},
\code{\link{is_sha256}()}
}
\concept{hash detectors}
