% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-03-get-stats.R
\name{is_missing}
\alias{is_missing}
\title{Check missing data}
\usage{
is_missing(x)
}
\arguments{
\item{x}{vector}
}
\value{
number of missing values in `x`
}
\description{
internal function for get_properties to return amount of missing data
}
\details{
internal function for \link{get_properties}
}
\keyword{internal}
