% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-datetime.R
\name{is_month_day}
\alias{is_month_day}
\title{Is it a Month-Day value (no year)?}
\usage{
is_month_day(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for `Month D` (e.g. `February 6`).
}
\description{
Is it a Month-Day value (no year)?
}
\examples{
is_month_day(c("February 6", "February 1987", NA))
}
\seealso{
Other datetime detectors: 
\code{\link{is_abbr_month_date}()},
\code{\link{is_clock_time}()},
\code{\link{is_date_range}()},
\code{\link{is_hijri_date}()},
\code{\link{is_iso_date}()},
\code{\link{is_iso_datetime}()},
\code{\link{is_long_date}()},
\code{\link{is_month_year}()},
\code{\link{is_rfc2822_datetime}()},
\code{\link{is_us_date}()},
\code{\link{is_us_datetime}()}
}
\concept{datetime detectors}
