% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-gov.R
\name{is_naics}
\alias{is_naics}
\title{Is it a NAICS industry code?}
\usage{
is_naics(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a 2-6 digit code present in the NAICS
  code list at any level (`.naics`, rolled up from the 2022 NAICS index).
}
\description{
Is it a NAICS industry code?
}
\examples{
is_naics(c("11", "541511", "99", NA))
}
\seealso{
Other gov detectors: 
\code{\link{is_chinese_resident_id}()},
\code{\link{is_dea}()},
\code{\link{is_ein}()},
\code{\link{is_ssn}()},
\code{\link{is_uei}()}
}
\concept{gov detectors}
