% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-phone.R
\name{is_phone}
\alias{is_phone}
\title{Is it a US / NANP phone number?}
\usage{
is_phone(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for common US formats such as `510-577-3201`,
  `(309) 647-1820`, `1-650-384-9911`, `650 231 1123`.
}
\description{
Is it a US / NANP phone number?
}
\examples{
is_phone(c("(309) 647-1820", "510-577-3201", "12345", NA))
}
\seealso{
Other phone detectors: 
\code{\link{is_intl_phone}()},
\code{\link{is_msisdn}()}
}
\concept{phone detectors}
