% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-color.R
\name{is_rgb_color}
\alias{is_rgb_color}
\title{Is it an RGB triple?}
\usage{
is_rgb_color(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for three comma-separated integers 0-255.
}
\description{
Is it an RGB triple?
}
\examples{
is_rgb_color(c("220,220,220", "128, 64, 32", "300,0,0", NA))
}
\seealso{
Other color detectors: 
\code{\link{is_cmyk_color}()},
\code{\link{is_hex_color}()},
\code{\link{is_hsl_color}()}
}
\concept{color detectors}
