% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-codes.R
\name{is_roman}
\alias{is_roman}
\title{Is it a Roman numeral?}
\usage{
is_roman(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a well-formed uppercase Roman numeral.
}
\description{
Is it a Roman numeral?
}
\examples{
is_roman(c("XIV", "MMXXIV", "IIII", "", NA))
}
\seealso{
Other code detectors: 
\code{\link{is_ca_postal_code}()},
\code{\link{is_country_code}()},
\code{\link{is_http_status}()},
\code{\link{is_uk_postcode}()},
\code{\link{is_unix_time}()},
\code{\link{is_us_state}()},
\code{\link{is_zip_code}()}
}
\concept{code detectors}
