% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-hash.R
\name{is_sha1}
\alias{is_sha1}
\title{Is it a SHA-1 digest?}
\usage{
is_sha1(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a 40-character hex string.
}
\description{
Is it a SHA-1 digest?
}
\examples{
is_sha1(c("da39a3ee5e6b4b0d3255bfef95601890afd80709", NA))
}
\seealso{
Other hash detectors: 
\code{\link{is_guid}()},
\code{\link{is_md5}()},
\code{\link{is_sha256}()}
}
\concept{hash detectors}
