% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-hash.R
\name{is_sha256}
\alias{is_sha256}
\title{Is it a SHA-256 digest?}
\usage{
is_sha256(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a 64-character hex string.
}
\description{
Is it a SHA-256 digest?
}
\examples{
is_sha256(c(paste(rep("a", 64), collapse = ""), NA))
}
\seealso{
Other hash detectors: 
\code{\link{is_guid}()},
\code{\link{is_md5}()},
\code{\link{is_sha1}()}
}
\concept{hash detectors}
