% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-address.R
\name{is_street_address}
\alias{is_street_address}
\title{Is it a street address?}
\usage{
is_street_address(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for `<number> <name> <street-type>`.
}
\description{
Is it a street address?
}
\examples{
is_street_address(c("2325 140th St", "1225 S Walnut St", "Bloomington", NA))
}
\seealso{
Other address detectors: 
\code{\link{is_address}()},
\code{\link{is_city_name}()},
\code{\link{is_city_state}()},
\code{\link{is_po_box}()},
\code{\link{is_state_zip}()}
}
\concept{address detectors}
