% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-codes.R
\name{is_uk_postcode}
\alias{is_uk_postcode}
\title{Is it a UK postcode?}
\usage{
is_uk_postcode(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a UK postcode like `SW1A 1AA`.
}
\description{
Is it a UK postcode?
}
\examples{
is_uk_postcode(c("SW1A 1AA", "IV40 8AJ", "12345", NA))
}
\seealso{
Other code detectors: 
\code{\link{is_ca_postal_code}()},
\code{\link{is_country_code}()},
\code{\link{is_http_status}()},
\code{\link{is_roman}()},
\code{\link{is_unix_time}()},
\code{\link{is_us_state}()},
\code{\link{is_zip_code}()}
}
\concept{code detectors}
