% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-scibio.R
\name{is_uniprot_id}
\alias{is_uniprot_id}
\title{Is it a UniProt accession?}
\usage{
is_uniprot_id(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a UniProtKB accession pattern.
}
\description{
Is it a UniProt accession?
}
\examples{
is_uniprot_id(c("P33667", "Q9BWD1", "ZZ", NA))
}
\seealso{
Other scibio detectors: 
\code{\link{is_asin}()},
\code{\link{is_bibcode}()},
\code{\link{is_ensembl_id}()},
\code{\link{is_icd10_code}()},
\code{\link{is_icd9_code}()},
\code{\link{is_snp_id}()}
}
\concept{scibio detectors}
