% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-codes.R
\name{is_us_state}
\alias{is_us_state}
\title{Is it a US state abbreviation?}
\usage{
is_us_state(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for a USPS two-letter state code (plus `DC`).
}
\description{
Is it a US state abbreviation?
}
\examples{
is_us_state(c("WA", "dc", "ZZ", NA))
}
\seealso{
Other code detectors: 
\code{\link{is_ca_postal_code}()},
\code{\link{is_country_code}()},
\code{\link{is_http_status}()},
\code{\link{is_roman}()},
\code{\link{is_uk_postcode}()},
\code{\link{is_unix_time}()},
\code{\link{is_zip_code}()}
}
\concept{code detectors}
