% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-codes.R
\name{is_zip_code}
\alias{is_zip_code}
\title{Is it a US ZIP code?}
\usage{
is_zip_code(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for `NNNNN` or `NNNNN-NNNN`.
}
\description{
Is it a US ZIP code?
}
\examples{
is_zip_code(c("98052", "98052-6399", "1234", NA))
}
\seealso{
Other code detectors: 
\code{\link{is_ca_postal_code}()},
\code{\link{is_country_code}()},
\code{\link{is_http_status}()},
\code{\link{is_roman}()},
\code{\link{is_uk_postcode}()},
\code{\link{is_unix_time}()},
\code{\link{is_us_state}()}
}
\concept{code detectors}
