% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{load_dgf}
\alias{load_dgf}
\title{Load DGF from DGF Excel Sheet}
\usage{
load_dgf(filename = "DGF.xlsx")
}
\arguments{
\item{filename}{A character string specifying the name of the Excel file to read. Defaults to `"DGF.xlsx"`.}
}
\value{
A data frame containing the data from the "DGF" sheet of the specified Excel file.
}
\description{
This function reads DGF from an Excel file and returns the contents of the "DGF" sheet as a data frame.
}
