% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-03-get-stats.R
\name{most_common_val}
\alias{most_common_val}
\title{Most common value}
\usage{
most_common_val(x)
}
\arguments{
\item{x}{vector}
}
\value{
most common value in `x`
}
\description{
Internal function for \link{get_properties} to find most common value.
}
\details{
internal function for \link{get_properties}
}
\keyword{internal}
