% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-05-apply-vname-aliases.R
\name{parse_nm}
\alias{parse_nm}
\title{Split a delimited alias string into unique names}
\usage{
parse_nm(x)
}
\arguments{
\item{x}{A single character string containing one or more names separated
by `";;"`.}
}
\value{
A character vector of unique, whitespace-trimmed names.
}
\description{
Splits a `";;"`-delimited string of variable-name aliases into a trimmed
vector of unique names.
}
