% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-00-BUILD-RG-DICT.R
\name{paste_properties}
\alias{paste_properties}
\title{Print Properties of a Variable into RG}
\usage{
paste_properties(VNAME, LABEL = "PROPERTIES")
}
\arguments{
\item{VNAME}{A character string specifying the name of the variable whose properties are to be displayed.}

\item{LABEL}{A character string (from layout.TYPE object) for the section title. Defaults to `"PROPERTIES"`.}
}
\value{
This function does not return a value; it prints the formatted table to the RG as html code
}
\description{
Read rg_properties column of DGF and print properties in a table in the RG
}
\details{
The function is an internal function of `create_div` in R/03-01-create-sections.R.
}
