% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-00-BUILD-RG-DICT.R
\name{paste_stats_log}
\alias{paste_stats_log}
\title{Print Statistic of a Logical Variable into RG}
\usage{
paste_stats_log(VNAME, LABEL = "STATS")
}
\arguments{
\item{VNAME}{A character string specifying the name of the variable whose properties are to be displayed.}

\item{LABEL}{A character string (from layout.TYPE object) for the section title. Defaults to `"STATS"`.}
}
\value{
This function does not return a value; it prints the formatted table to the RG as html code
}
\description{
Read rg_stats column of DGF and print statistics in a table in the RG
}
\details{
The function is an internal function of `create_div` in R/03-01-create-sections.R.
}
