% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-02-treemap.R
\name{paste_treemap}
\alias{paste_treemap}
\title{Print Graphics of a Factor Variable into RG}
\usage{
paste_treemap(VNAME, LABEL = "MOST COMMON VALUES")
}
\arguments{
\item{VNAME}{A character string specifying the name of the variable whose properties are to be displayed.}

\item{LABEL}{A character string (from layout.TYPE object) for the section title. Defaults to `"MOST COMMON VALUES"`.}
}
\value{
This function does not return a value; it prints the formatted treemap to the RG
}
\description{
Read rg_graphics column of DGF and print treemap in the RG
}
\details{
The function is an internal function of `create_div` in R/03-01-create-sections.R.
}
