% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-02-data-type-factors.R
\name{recast_factors}
\alias{recast_factors}
\title{Convert guessed factor columns to factors}
\usage{
recast_factors(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
The input data frame with guessed-categorical columns coerced to
  factors.
}
\description{
Uses [is_factor_df()] to identify columns that should be treated as
categorical and converts each of them to a factor, leaving all other
columns unchanged.
}
\seealso{
[is_factor_df()]
}
