% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-05-apply-vname-aliases.R
\name{rename_all}
\alias{rename_all}
\title{Rename data frame columns using an alias mapping}
\usage{
rename_all(df, old, new)
}
\arguments{
\item{df}{A data frame whose columns should be renamed.}

\item{old}{A character vector of aliases (each element optionally
`";;"`-delimited) identifying columns to rename.}

\item{new}{A character vector of replacement names, parallel to `old`.}
}
\value{
The data frame with renamed columns.
}
\description{
Renames the columns of `df` by applying [replace_name()] for each
`old`/`new` pair, where each `old` entry may be a `";;"`-delimited set of
aliases mapping to a single `new` name.
}
\seealso{
[replace_name()]
}
