% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-02-wordcloud.R
\name{v_to_wordcloud}
\alias{v_to_wordcloud}
\title{Print Graphics of a character Variable into RG}
\usage{
v_to_wordcloud(VNAME, LABEL = "WORD CLOUD")
}
\arguments{
\item{VNAME}{A character string specifying the name of the variable whose properties are to be displayed.}

\item{LABEL}{A character string (from layout.TYPE object) for the section title. Defaults to `"Word Cloud"`.}
}
\value{
This function does not return a value; it prints the formatted word cloud to the RG
}
\description{
Read rg_graphics column of DGF and print word cloud in the RG
}
\details{
The function is an internal function of `create_div` in R/03-01-create-sections.R.
}
