% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-00-INGEST-RAW.R
\name{apply_name_aliases}
\alias{apply_name_aliases}
\title{FUNCTION_TITLE}
\usage{
apply_name_aliases(df, dgf)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{dgf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[crosswalkr]{renamefrom}}
}
