% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-00-INGEST-RAW.R
\name{apply_raw_convert_fx}
\alias{apply_raw_convert_fx}
\title{FUNCTION_TITLE}
\usage{
apply_raw_convert_fx(df, dgf)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{dgf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate_all}}
}
