% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-00-CREATE-DGF.R
\name{create_dgf}
\alias{create_dgf}
\title{create DGF file}
\usage{
create_dgf(
  df,
  vtypes = NULL,
  use.df.types = FALSE,
  guess.factors = TRUE,
  guess.dates = FALSE,
  dd = NULL,
  vname = NULL,
  vlabel = NULL,
  vdesc = NULL,
  vtype = NULL,
  vconvert = NULL,
  vformat = NULL,
  vname_alias = NULL,
  vscope = NULL,
  vloc = NULL,
  keep.dd.cols = NULL,
  preview_dd = F,
  preview_dp = F,
  file = "DGF"
)
}
\description{
some description text
}
\details{
some additional details
}
