% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-01-create-sections.R
\name{create_div}
\alias{create_div}
\title{Create all non-level-1 div section in markdown}
\usage{
create_div(div.num = "div2", all.layouts, xx)
}
\arguments{
\item{all.layouts}{output of \link{get_design}}

\item{xx}{defined as a global variable in \link{create_section}.}

\item{x}{Character string, the label or variable name to include
in the header. Defaults to `"vname"`.}
}
\value{
No return value, called for side effects (writes markdown text).
}
\description{
Writes a formatted `div` block to the quarto document. Format of the `div`
block is determined by the `all.layouts` table.
}
\details{
If div name of `x` is not included in `all.layouts`, nothing will be
printed for that div in the quarto document.
}
