% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{first_n}
\alias{first_n}
\title{Extract and Format the First N Unique Values from a Vector}
\usage{
first_n(x, n = 5)
}
\arguments{
\item{x}{A vector of any type (numeric, character, factor, or Date).}

\item{n}{An integer specifying the number of unique values to return. Defaults to 5.}
}
\value{
A formatted character string containing up to the first N unique values from `x`,
separated by " ;; \n". If `x` is numeric, values are rounded to three decimal places.
If `x` contains factors or dates, they are converted to character strings.
}
\description{
This function extracts the first N unique, non-missing values from a vector, formats them appropriately,
and returns them as a truncated character string.
}
\examples{
first_n(c("apple", "banana", "cherry", "date", "elderberry", "fig"), 3)
first_n(c(1.234567, 2.345678, 3.456789), 2)
first_n(as.Date(c("2021-01-01", "2021-02-01", "2021-03-01")), 2)

}
\seealso{
\code{\link[stringr]{str_trunc}}
}
