% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{get_class_df}
\alias{get_class_df}
\title{FUNCTION_TITLE}
\usage{
get_class_df(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
