% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-03-get-graphics.R
\name{get_graphics}
\alias{get_graphics}
\title{Generate Graphics Based on Variable Type}
\usage{
get_graphics(VNAME, df, VCLASS)
}
\arguments{
\item{VNAME}{A character string specifying the name of the variable in the data frame. (From vname column of the DGF)}

\item{df}{A data frame containing the variable to be visualized.}

\item{VCLASS}{A character string indicating the class of the variable (e.g., "numeric", "factor", "logical", or other).}
}
\value{
Data table as JSON text string needed to store in the DGF to later generate the appropriate plot when rendering the RG.
}
\description{
Generates needed graphics for rg_graphics column of the DGF. Used inside R/01-00-CREATE-DGF.R
}
\details{
The function calls different helper functions based on `VCLASS`:
\itemize{
  \item `"numeric"`: Calls `get_graphics_num(VNAME, df)`
  \item `"factor"`: Calls `get_graphics_fact(VNAME, df)`
  \item `"logical"`: Calls `get_graphics_log(VNAME, df)`
  \item Other classes: Calls `get_graphics_chr(VNAME, df)`
}
}
