% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{get_levels}
\alias{get_levels}
\title{FUNCTION_TITLE}
\usage{
get_levels(f)
}
\arguments{
\item{f}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
