% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-00-INGEST-RAW.R
\name{ingest_raw}
\alias{ingest_raw}
\title{FUNCTION_TITLE}
\usage{
ingest_raw(df, path = NULL)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{path}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
