% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-datetime.R
\name{is_date_range}
\alias{is_date_range}
\title{Is it a date range?}
\usage{
is_date_range(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for two ISO or US dates joined by `to`/`-`.
}
\description{
Is it a date range?
}
\examples{
is_date_range(c("07/21/16 to 08/19/16", "1972-01-01 - 1972-07-01", NA))
}
\seealso{
Other datetime detectors: 
\code{\link{is_abbr_month_date}()},
\code{\link{is_clock_time}()},
\code{\link{is_hijri_date}()},
\code{\link{is_iso_date}()},
\code{\link{is_iso_datetime}()},
\code{\link{is_long_date}()},
\code{\link{is_month_day}()},
\code{\link{is_month_year}()},
\code{\link{is_rfc2822_datetime}()},
\code{\link{is_us_date}()},
\code{\link{is_us_datetime}()},
\code{\link{is_yyyymm}()}
}
\concept{datetime detectors}
