% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-gov.R
\name{is_ein}
\alias{is_ein}
\title{Is it a US Employer Identification Number (EIN)?}
\usage{
is_ein(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; `TRUE` for `dd-ddddddd`.
}
\description{
Is it a US Employer Identification Number (EIN)?
}
\examples{
is_ein(c("15-4464349", "154464349", NA))
}
\seealso{
Other gov detectors: 
\code{\link{is_chinese_resident_id}()},
\code{\link{is_dea}()},
\code{\link{is_naics}()},
\code{\link{is_ssn}()},
\code{\link{is_uei}()}
}
\concept{gov detectors}
