% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-02-data-type-factors.R
\name{is_factor}
\alias{is_factor}
\title{FUNCTION_TITLE}
\usage{
is_factor(
  x,
  unique.p = 0.05,
  unique.n = 100,
  top.n = 25,
  max.variance = 2,
  b.to.f = FALSE
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{unique.p}{PARAM_DESCRIPTION, Default: 0.05}

\item{unique.n}{PARAM_DESCRIPTION, Default: 100}

\item{top.n}{PARAM_DESCRIPTION, Default: 25}

\item{max.variance}{PARAM_DESCRIPTION, Default: 2}

\item{b.to.f}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[DescTools]{IsDate}}
 \code{\link[hablar]{retype}}
}
