% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{jsonify_f}
\alias{jsonify_f}
\title{FUNCTION_TITLE}
\usage{
jsonify_f(f)
}
\arguments{
\item{f}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
}
