% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-00-INGEST-RAW.R
\name{parse_functions}
\alias{parse_functions}
\title{FUNCTION_TITLE}
\usage{
parse_functions(raw_convert)
}
\arguments{
\item{raw_convert}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
