% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-05-apply-vname-aliases.R
\name{parse_nm}
\alias{parse_nm}
\title{FUNCTION_TITLE}
\usage{
parse_nm(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
