% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-02-booleplot.R
\name{paste_booleplot}
\alias{paste_booleplot}
\title{Print Graphics of a Logical Variable into RG}
\usage{
paste_booleplot(VNAME, LABEL = "BAR PLOT")
}
\arguments{
\item{VNAME}{A character string specifying the name of the variable whose properties are to be displayed.}

\item{LABEL}{A character string (from layout.TYPE object) for the section title. Defaults to `"BAR PLOT"`.}
}
\value{
This function does not return a value; it prints the formatted Booleplot to the RG
}
\description{
Read rg_graphics column of DGF and print Booleplot in the RG
}
\details{
The function is an internal function of `create_div` in R/03-01-create-sections.R.
}
