% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-02-histogram.R
\name{paste_histogram}
\alias{paste_histogram}
\title{Print Graphics of a Numeric Variable into RG}
\usage{
paste_histogram(VNAME, LABEL = "HISTOGRAM")
}
\arguments{
\item{VNAME}{A character string specifying the name of the variable whose properties are to be displayed.}

\item{LABEL}{A character string (from layout.TYPE object) for the section title. Defaults to `"HISTOGRAM"`.}
}
\value{
This function does not return a value; it prints the formatted histogram to the RG
}
\description{
Read rg_graphics column of DGF and print histogram in the RG
}
\details{
The function is an internal function of `create_div` in R/03-01-create-sections.R.
}
