% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-00-BUILD-RG-DICT.R
\name{paste_preview_chr}
\alias{paste_preview_chr}
\title{Print Preview of a Character Variable into RG}
\usage{
paste_preview_chr(VNAME, LABEL = "PREVIEW")
}
\arguments{
\item{VNAME}{A character string specifying the name of the variable whose properties are to be displayed.}

\item{LABEL}{A character string (from layout.TYPE object) for the section title. Defaults to `"PREVIEW"`.}
}
\value{
This function does not return a value; it prints the formatted table to the RG as html code
}
\description{
Read rg_preview column of DGF and print preview in a table in the RG
}
\details{
The function is an internal function of `create_div` in R/03-01-create-sections.R.
}
