% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-02-data-type-factors.R
\name{recast_factors}
\alias{recast_factors}
\title{FUNCTION_TITLE}
\usage{
recast_factors(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
