% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-05-apply-vname-aliases.R
\name{rename_all}
\alias{rename_all}
\title{FUNCTION_TITLE}
\usage{
rename_all(df, old, new)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{old}{PARAM_DESCRIPTION}

\item{new}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
