% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-05-apply-vname-aliases.R
\name{replace_name}
\alias{replace_name}
\title{FUNCTION_TITLE}
\usage{
replace_name(nm, x, y)
}
\arguments{
\item{nm}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
