% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-04-save-to-excel.R
\name{save_to_excel}
\alias{save_to_excel}
\title{Saves DGF to excel}
\usage{
save_to_excel(df, filename = "DGF.xlsx")
}
\description{
Write the DGF to excel.
}
\details{
used inside of 01-00-CREATE-DGF.R
}
