% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-04-validate-json.R
\name{show_invalid}
\alias{show_invalid}
\title{FUNCTION_TITLE}
\usage{
show_invalid(v)
}
\arguments{
\item{v}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
