% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-04-validate-json.R
\name{validate_json}
\alias{validate_json}
\title{FUNCTION_TITLE}
\usage{
validate_json(v)
}
\arguments{
\item{v}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[jsonlite]{validate}}
}
